# cop-sys-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For COP-SYSTEM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/case": {
            "get": {
                "tags": [
                    "Case"
                ],
                "summary": "Get all cases",
                "operationId": "533fc7ed6ca8c332730e57c0ef4f470b",
                "responses": {
                    "200": {
                        "description": "Cases list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CasesResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Case"
                ],
                "summary": "Add a new case",
                "operationId": "60314d946d91725582b5c6be8c37d5bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Case added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CasesResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/case/{case}": {
            "get": {
                "tags": [
                    "Case"
                ],
                "summary": "View a case by ID",
                "operationId": "bf5cc293b1375c062bd8f2a1e6070bdd",
                "parameters": [
                    {
                        "name": "case",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Case retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CasesResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Case not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Case"
                ],
                "summary": "Update an existing case",
                "operationId": "08f4a4a93e8b85e2bb6d712f8f2467ef",
                "parameters": [
                    {
                        "name": "case",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Case updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CasesResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Case not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Case"
                ],
                "summary": "Remove a case",
                "operationId": "659db5ec41158f47c334685dad2e7e1f",
                "parameters": [
                    {
                        "name": "case",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Case deleted successfully"
                    },
                    "404": {
                        "description": "Case not found"
                    }
                }
            }
        },
        "/api/v1/case/investigator/{personnel_id}": {
            "get": {
                "tags": [
                    "Case"
                ],
                "summary": "Get cases by investigator ID",
                "operationId": "0863267972f0b7cf481fee3f8ac93424",
                "parameters": [
                    {
                        "name": "personnel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cases retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CasesResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cases not found"
                    }
                }
            }
        },
        "/api/v1/conversation/admin": {
            "get": {
                "tags": [
                    "Conversation"
                ],
                "summary": "Get the conversation with admins",
                "operationId": "bb65c663e2102ba417a23df6760f63ff",
                "responses": {
                    "200": {
                        "description": "Admin conversation retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConversationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Admin conversation not found"
                    }
                }
            }
        },
        "/api/v1/department/{department_id}/personnels": {
            "get": {
                "tags": [
                    "Department Personnel"
                ],
                "summary": "Get personnels by department",
                "operationId": "8796ab7bed07e3d73fb4a8c7535cf9fb",
                "parameters": [
                    {
                        "name": "department_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personnels retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DepartmentPersonnelResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Department not found"
                    }
                }
            }
        },
        "/api/v1/personnel/{personnel_id}/departments": {
            "get": {
                "tags": [
                    "Department Personnel"
                ],
                "summary": "Get departments by personnel",
                "operationId": "a557f705e71c067ec9b0373a17894699",
                "parameters": [
                    {
                        "name": "personnel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Departments retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DepartmentPersonnelResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Personnel not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PersonnelsResource": {
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "rank": {
                        "type": "string"
                    },
                    "badge_number": {
                        "type": "string"
                    },
                    "hire_date": {
                        "type": "string"
                    },
                    "profile_image": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CreateCaseRequest": {
                "required": [
                    "status",
                    "open_date",
                    "investigator_id",
                    "incidents_id"
                ],
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "open_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "close_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "investigator_id": {
                        "type": "string"
                    },
                    "incidents_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateCaseRequest": {
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "open_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "close_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "investigator_id": {
                        "type": "string"
                    },
                    "incidents_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CasesResource": {
                "properties": {
                    "case_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "open_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "close_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "investigator_id": {
                        "type": "string"
                    },
                    "incidents_id": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ConversationResource": {
                "properties": {
                    "conversation_id": {
                        "type": "string"
                    },
                    "conversation_name": {
                        "type": "string"
                    },
                    "conversation_picture": {
                        "type": "string"
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "DepartmentsResource": {
                "properties": {
                    "department_id": {
                        "type": "integer"
                    },
                    "department_name": {
                        "type": "string"
                    },
                    "department_logo": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DepartmentPersonnelResource": {
                "properties": {
                    "department_id": {
                        "type": "integer"
                    },
                    "personnel_id": {
                        "type": "integer"
                    },
                    "personnels": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PersonnelsResource"
                        }
                    },
                    "departments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DepartmentsResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}